RGA_ADK/requirements.txt
google-cloud-firestore==2.14.0

# エンベディング（ベクトル変換）
sentence-transformers==2.3.1
torch==2.1.2

# 数値計算
numpy==1.24.3

# 進捗バー（オプション）
tqdm==4.66.1

# LLM統合（後で使う）
openai==1.12.0
# または
# anthropic==0.18.1

# MCPサーバー化用（後で使う）
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# ユーティリティ
python-dotenv==1.0.1